Refactor: Moved ZoneMenu to components folder and updated imports
